+++ features/leave/components/confirmation-dialog.tsx
@@ -9,7 +9,7 @@ import {
   AlertDialogFooter,
   AlertDialogHeader,
   AlertDialogTitle,
-} from '@/components/ui/alert-dialog';
+} from '@/components/shadcn/alert-dialog';
 
 interface ConfirmationDialogProps {
   open: boolean;
