+++ features/leave/components/empty-state.tsx
@@ -1,5 +1,5 @@
-import { Card, CardContent } from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
+import { Card, CardContent } from '@/components/shadcn/card';
+import { Button } from '@/components/shadcn/button';
 import { LucideIcon } from 'lucide-react';
 
 interface EmptyStateProps {
